working indirect build
3/11/2021

+++ core/lib/brdfs/Lambert.cpp
@@ -18,19 +18,19 @@ namespace Tracer {
 	namespace BRDF {
 
 		Vector3 Lambert::Sampler(double r1, double r2) {
-			float sinTheta = sqrt(1 - r1 * r1);
-			float phi = 2 * M_PI * r2;
-			float x = sinTheta * cos(phi);
-			float z = sinTheta * sin(phi);
+			double sinTheta = sqrt(1.0 - r1 * r1);
+			double phi = 2.0 * M_PI * r2;
+			double x = sinTheta * cos(phi);
+			double z = sinTheta * sin(phi);
 
 			return Vector3(x, r1, z);
 		}
 
 		void Lambert::CreateCoordinateSystem(const Vector3& N, Vector3& Nt, Vector3& Nb) {
 			if (std::abs(N.x) > std::abs(N.y))
-				Nt = Vector3(N.z, 0, -N.x) / sqrt(N.x * N.x + N.z * N.z);
+				Nt = Vector3(N.z, 0.0, -N.x) / sqrt(N.x * N.x + N.z * N.z);
 			else
-				Nt = Vector3(0, -N.z, N.y) / sqrt(N.y * N.y + N.z * N.z);
+				Nt = Vector3(0.0, -N.z, N.y) / sqrt(N.y * N.y + N.z * N.z);
 			Nb = N.cross(Nt);
 		}
